Support repo's Appstream data download and install
Repositories can provide Appstream data for the packages they contain.
This Appstream data is consumed by applications like gnome-software or
KDE Discover, thus the users can see the packages (apps) in them.

This is to be in pair with PackageKit, which does download and install
the repo's Appstream data.

As this adds a dependency on the `appstream` library, there is also
a CMake option WITH_APPSTREAM to be able to turn the support off.
The support is enabled by default.